requirements.txt
hikari
azure-storage-blob